backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
celery==5.3.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
dj-rest-auth==4.0.1
Django==4.2.3
django-cors-headers==4.2.0
django-silk==5.0.3
djangorestframework==3.14.0
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
sqlparse==0.4.4
gunicorn==20.1.0
mysqlclient==2.0.3
